+++ R/utils.r
@@ -103,3 +103,5 @@ skip_on_appveyor <- function() {
 
   skip("On Appveyor")
 }
+
+trim_whitespace <- function(text) gsub("^\\s*|\\s*$", "", text, perl = TRUE)